parse_fastq/requirements.txt
numpy==1.11.0
biopython==1.67
tqdm==4.7.4
